more robust registration of post type capabilities
The prior version did not handle post types with custom values
for the 'capabilities' property, and required pluralization
by string concatenation in the caller.

The new version uses the actually capabilities assigned
to the post type, and can use the real post ID as an argument.

I've maintained the prior version for backwards compatibility,
in the event that the method is called with the capability
suffix rather than the post type name.